Added an option for a flash block image
Browsing to a flash file which gets blocked with a plain text block page seems
to take out the flash plugin and then the browser reasonably successfully
(using firefox at any rate). Added an optional flash block page which works
the same way as the image blocking to stop the browser crashing.



git-svn-id: https://svn.code.sf.net/p/dansguardian/code/trunk@1055 c6e6bffc-5bc6-4360-8735-ace3d95093af